fix(engines): ensure correct required node version
Closes #237

BREAKING CHANGE:

Changed `engines.node` from `>=16` to `>=18.18.0`.

`minimatch@10` was updated in `syncpack@12.4.0` which required `node@>=20` and
caused issues in some projects.

This change adds a local linter to verify that the node engines of syncpack's
dependencies all satisfy syncpack's own node engine. Using this linter found
that the minimum node engine of all of syncpack's dependencies is v18.18.0.